# storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "Brokers Web API",
        "description": "Esta es la documentación para la API que manejará la información de las propiedades o inmuebles  You can find\n out more about Swagger at\n[http://swagger.io](http://swagger.io) or on\n[irc.freenode.net, #swagger](http://swagger.io/irc/).",
        "termsOfService": "http://swagger.io/terms/",
        "contact": {
            "email": "yocumo1998@gmail.com"
        },
        "license": {
            "name": "Apache 2.0",
            "url": "http://www.apache.org/licenses/LICENSE-2.0.html"
        },
        "version": "1.0.0"
    },
    "paths": {
...
        "/api/cosigners": {
            "get": {
                "tags": [
                    "Cosigner"
                ],
                "summary": "Obtener todos los codeudores",
                "operationId": "cosigners",
                "responses": {
                    "default": {
                        "description": "successful operation"
                    }
                },
                "security": [
                    {
                        "bearer": []
                    }
                ]
            }
        },
        "/api/cosigners/{id}": {
            "get": {
                "tags": [
                    "Cosigner"
                ],
                "summary": "Obtener un codeudor",
                "operationId": "cosigner/{id}",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "ID del codeudor",
                        "required": true,
                        "schema": {
                            "type": "integer",
                            "format": "int64"
                        }
                    }
                ],
                "responses": {
                    "default": {
                        "description": "successful operation"
                    }
                },
                "security": [
                    {
                        "bearer": []
                    }
                ]
            },
            "delete": {
                "tags": [
                    "Cosigner"
                ],
                "summary": "Eliminar un codeudor",
                "operationId": "cosigner",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "ID del codeudor",
                        "required": true,
                        "schema": {
                            "type": "integer",
                            "format": "int64"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "successful operation"
                    },
                    "404": {
                        "description": "Codeudor no encontrado"
                    }
                },
                "security": [
                    {
                        "bearer": []
                    }
                ]
            }
        },
...
    },
    "components": {
...
        "securitySchemes": {
            "bearer": {
                "type": "apiKey",
                "description": "Ingrese el token en formato (Bearer <token>)",
                "name": "api_key",
                "in": "header"
            }
        }
    },
    "security": {
        "bearer": {
            "type": "apiKey",
            "description": "Ingrese el token en formato (Bearer <token>)",
            "name": "api_key",
            "in": "header"
        }
    }
}